requirements.txt
fastapi[standard]
python-dotenv
pytest
google-genai
